Клиентское кэширование с react-query
