+++ src/constants/queryKeys.js
@@ -1,0 +1,3 @@
+export const QUERY_KEYS = {
+    AIRTABLE_DATA: ['airtableData']
+}
