+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { createRoot } from 'react-dom/client'; // Импортируем createRoot
-import App from './App';
+import { App } from './App';
 import './index.css';
 
 const container = document.getElementById('root');
